+++ CHANGELOG.md
@@ -5,6 +5,7 @@ CHANGE LOG
  - Fix PSR4 autoloading
  - Add PHP interfaces for scripts, languages, locales, etc
  - Add several missing plural rules
+ - Add Locale::httpAcceptLanguage() to negotiate languages
 
 ## 1.3.2 (2015-03-23)
  - Gettext detection fails on 32 bit builds